add root exceptions module and move participant exceptions there
