+++ gittip/elsewhere/__init__.py
@@ -6,9 +6,9 @@ from aspen.utils import typecheck
 from psycopg2 import IntegrityError
 
 import gittip
+from gittip.exceptions import ProblemChangingUsername
 from gittip.security.user import User
 from gittip.models.participant import Participant, reserve_a_random_username
-from gittip.models.participant import ProblemChangingUsername
 
 
 ACTIONS = [u'opt-in', u'connect', u'lock', u'unlock']
