+++ www/%username/username.json.spt
@@ -1,6 +1,6 @@
 from aspen import Response, log_dammit
+from gittip.exceptions import ProblemChangingUsername
 from gittip.models.participant import Participant
-from gittip.models.participant import ProblemChangingUsername
 
 
 [-----------------------------------------------------------------------------]
